fix(gateway): address some more feedback
